Refactor static-y ArgumentTransformer methods into new BoundMethod
Previously a lot of behavior was independently consolidated in
ArgumentTransformer; this led to a lot of repeated code as it was a
"global" class (it lives for the entire lifetime of a Frame instance).
The new BoundMethod class allows us to make many improvements &
optimizations; for example, we now pre-compute each parameter's
provider at injection time, saving us doing it on every command
execution.

Version bumped to 1.4.0-PRE; breaking changes have been made.